fix(uci): parse every `go` argument at upstream's own width
Third of the fourth rfish sweep (its dcc5a14). `is >> field` fails when the text
does not fit the FIELD's type, and upstream turns that failbit into a critical
error on the very next line (uci.cpp:229-230) -- so the declared width is
observable from a GUI. mcfish parsed everything through one `strtol`, which is too
wide for the five `int` fields and too narrow for `nodes`:

  go depth 3000000000                 upstream: CRITICAL ERROR   mcfish: searched
  go mate 3000000000                  upstream: CRITICAL ERROR   mcfish: searched
  go movestogo 3000000000 wtime 1000  upstream: CRITICAL ERROR   mcfish: searched
  go perft 3000000000                 upstream: CRITICAL ERROR   mcfish: divided
  go movetime 99999999999999999999    upstream: CRITICAL ERROR   mcfish: searched
  go wtime 9223372036854775808        upstream: CRITICAL ERROR   mcfish: searched
  go nodes 999999999999999999999999   upstream: CRITICAL ERROR   mcfish: searched

Each key now parses at upstream's own type: `int` for depth/mate/movestogo/perft,
int64_t for the four clocks, `u64` for nodes -- the last read the way a C++ stream
reads one, where a leading minus is accepted and the magnitude wraps, so `go nodes
-1` stays a budget of UINT64_MAX on both sides. The bound IS the accept/reject
boundary, not something to clamp to, which is the same rule the option table
already follows for an out-of-range spin.

THE CLOCK FIELDS WERE ALSO TOO NARROW TO HOLD WHAT UPSTREAM ACCEPTS. LimitsType's
`time`, `inc` and `movetime` are `TimePoint` -- int64_t -- and SearchLimits held
them as `int`, so `go wtime 3000000000` (34 days, a clock upstream takes) was
truncated rather than refused. They are int64_t now; every consumer already cast
to TimePoint on the way into timeman.

Probed against the pinned oracle rather than inferred: 20 of 20 `go` argument
cases agree, seven of which diverged before. tools/cases/gowidth.uci pins both
sides of one boundary -- `go nodes 18446744073709551615 depth 1` searches,
`go depth 3000000000` is the error -- with a golden written from the oracle.

Gates (exit 0, each read from the gate itself): parity -- build, zone-check, fmt,
docs-lint, test 1544973 checks 0 failures under ASan+UBSan, signature 2508687
UNCHANGED (bench passes its limits through benchmark.c, not this parse),
simd-scalar, perft, golden 19 cases, tb; golden-audit 19 agree 0 differ 0 missing;
no gate skipped.

+++ src/engine/search/search.h
@@ -13,10 +13,14 @@
 #include "../board/types.h"
 
 typedef struct {
-    int depth;        // fixed-depth limit, 0 for none
-    int movetime_ms;  // fixed time per move, 0 for none
-    int time_ms[COLOR_NB];
-    int inc_ms[COLOR_NB];
+    int depth;  // fixed-depth limit, 0 for none
+    // The four clock fields are upstream's TimePoint (search.h LimitsType), which is
+    // int64_t and NOT int: `go wtime 3000000000` is a clock upstream accepts, and
+    // holding it in an int truncated it here. Declared as int64_t rather than as
+    // TimePoint so this header stays independent of timeman.h.
+    int64_t movetime_ms;  // fixed time per move, 0 for none
+    int64_t time_ms[COLOR_NB];
+    int64_t inc_ms[COLOR_NB];
     int moves_to_go;
     uint64_t nodes;  // node limit, 0 for none
     // `go mate N`: stop once the best score is a mate in at most N MOVES -- the
